cleanup KBC project as trait
