+++ tests/CleanupKbcProject.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\ProjectBackup\Tests;
+
+use Keboola\StorageApi\Client;
+use Keboola\StorageApi\Components;
+
+trait CleanupKbcProject
+{
+    protected Client $sapiClient;
+
+    private function cleanupKbcProject(): void
+    {
+        $components = new Components($this->sapiClient);
+        foreach ($components->listComponents() as $component) {
+            foreach ($component['configurations'] as $configuration) {
+                $components->deleteConfiguration($component['id'], $configuration['id']);
+
+                // delete configuration from trash
+                $components->deleteConfiguration($component['id'], $configuration['id']);
+            }
+        }
+
+        // drop linked buckets
+        foreach ($this->sapiClient->listBuckets() as $bucket) {
+            if (isset($bucket['sourceBucket'])) {
+                $this->sapiClient->dropBucket($bucket['id'], ['force' => true]);
+            }
+        }
+
+        foreach ($this->sapiClient->listBuckets() as $bucket) {
+            $this->sapiClient->dropBucket($bucket['id'], ['force' => true]);
+        }
+    }
+}
